*** empty log message ***

+++ lib/gnutls_handshake.c
@@ -2815,7 +2815,7 @@ _gnutls_remove_unwanted_ciphersuites (gnutls_session_t session,
   *
   **/
 void
-gnutls_handshake_set_max_packet_length (gnutls_session_t session, int max)
+gnutls_handshake_set_max_packet_length (gnutls_session_t session, size_t max)
 {
   session->internals.max_handshake_data_buffer_size = max;
 }
